fixup! feat(suite#4941): create getNonAsciiChars fn
