--- packages/utils/src/isAscii.ts
@@ -1,5 +1,0 @@
-export function isAscii(value?: string): boolean {
-    if (!value) return true;
-
-    return /^[\x00-\x7F]*$/.test(value);
-}
